Refactor frontend modules and remove bundle
