+++ index.html
@@ -64,6 +64,6 @@
             <div id="study-container" class="hidden"></div>
         </section>
     </div>
-    <script src="bundle.js" defer></script>
+    <script type="module" src="src/app.js"></script>
 </body>
 </html>
